% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_truth.R
\name{calc_truth_agecomp_yearly}
\alias{calc_truth_agecomp_yearly}
\title{Calculate "true" yearly age composition for a species}
\usage{
calc_truth_agecomp_yearly(truth_agecomp_monthly)
}
\arguments{
\item{truth_agecomp_monthly}{A tibble containing the monthly age composition
data for a species. Returned from \code{\link[=calc_truth_agecomp_monthly]{calc_truth_agecomp_monthly()}}.}
}
\value{
A tibble containing the yearly age composition (e.g., biomass, catch,
and weight) for the specified species.
}
\description{
This function calculates the yearly age composition for a species
by averaging the monthly age composition values.
}
\examples{
data(ewe_nwatlantic_base, package = "ecosystemdata")
data <- ewe_nwatlantic_base
truth_agecomp_yearly <- ecosystemom:::calc_truth_monthly(
  data,
  species_name = "menhaden",
  truth_type = "biomass"
) |>
  ecosystemom:::calc_truth_agecomp_monthly() |>
  ecosystemom:::calc_truth_agecomp_yearly()
}
\keyword{calc_truth}
