% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_truth.R
\name{calc_truth_index_monthly}
\alias{calc_truth_index_monthly}
\title{Calculate "true" monthly index for a species}
\usage{
calc_truth_index_monthly(truth_monthly)
}
\arguments{
\item{truth_monthly}{A tibble containing the monthly "truth" data for a species.
Returned from \code{\link[=calc_truth_monthly]{calc_truth_monthly()}}.}
}
\value{
A tibble containing the monthly index (e.g., biomass and catch) for
the specified species
}
\description{
This function calculates the monthly index for a species by
summing the values across all ages by month.
}
\examples{
data(ewe_ecosim_base_nwatlantic)
data <- ewe_ecosim_base_nwatlantic
truth_index_monthly <- ecosystemom:::calc_truth_monthly(
  data,
  species_name = "menhaden",
  truth_type = "biomass"
) |>
  ecosystemom:::calc_truth_index_monthly()
}
\keyword{calc_truth}
