% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_truth.R
\name{calc_truth_index_yearly}
\alias{calc_truth_index_yearly}
\title{Calculate "true" yearly index for a species}
\usage{
calc_truth_index_yearly(truth_index_monthly)
}
\arguments{
\item{truth_index_monthly}{A tibble containing the monthly index data for a species.
Returned from \code{\link[=calc_truth_index_monthly]{calc_truth_index_monthly()}}.}
}
\value{
A tibble containing the yearly index (e.g., biomass and catch) for
the specified species.
}
\description{
This function calculates the yearly index for a species by
averaging monthly index values.
}
\examples{
data(ewe_ecosim_base_nwatlantic)
data <- ewe_ecosim_base_nwatlantic
truth_index_yearly <- ecosystemom:::calc_truth_monthly(
  data,
  species_name = "menhaden",
  truth_type = "biomass"
) |>
  ecosystemom:::calc_truth_index_monthly() |>
  ecosystemom:::calc_truth_index_yearly()
}
\keyword{calc_truth}
