% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_truth.R
\name{calc_truth_monthly}
\alias{calc_truth_monthly}
\title{Caculate monthly "truth" for a species}
\usage{
calc_truth_monthly(data, species_name, truth_type)
}
\arguments{
\item{data}{A tibble containing ecosystem model outputs from {ecosystemdata}.}

\item{species_name}{A string specifying the species name to subset.}

\item{truth_type}{A string specifying the type of "truth" to calculate.}
}
\value{
A tibble containing the monthly "truth" data for the specified species.
}
\description{
This function calculates the monthly "truth" for a species by
filtering the data for the specified species and type
}
\examples{
data(ewe_nwatlantic_base, package = "ecosystemdata")
data <- ewe_nwatlantic_base
truth_monthly <- ecosystemom:::calc_truth_monthly(
  data,
  species_name = "menhaden",
  truth_type = "biomass"
)
}
\keyword{calc_truth}
