% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosystemom-package.R
\docType{package}
\name{ecosystemom-package}
\alias{ecosystemom}
\alias{ecosystemom-package}
\title{ecosystemom: Create sampled data from standardized ecosystem output}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Bai Li \email{bai.li@noaa.gov} (\href{https://orcid.org/0000-0002-8249-1442}{ORCID})

Authors:
\itemize{
  \item Bai Li \email{bai.li@noaa.gov} (\href{https://orcid.org/0000-0002-8249-1442}{ORCID})
  \item Kelli F. Johnson \email{kelli.johnson@noaa.gov} (\href{https://orcid.org/0000-0002-5149-451X}{ORCID})
}

}
\keyword{internal}
