% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_true_maturity.R
\name{estimate_true_maturity}
\alias{estimate_true_maturity}
\title{Estimate true maturity parameters from ecosystem model input}
\usage{
estimate_true_maturity(ages, spawning_proportion, functional_form = "logistic")
}
\arguments{
\item{ages}{A vector of ages.}

\item{spawning_proportion}{A numeric vector of spawning proportions (0 to 1)
for each functional group within a species. The values can be obtained from
Ecopath -> Input -> Basic input -> Edit Multi_Stanza Groups.}

\item{functional_form}{A character string specifying the curve
type. Currently supports "logistic".}
}
\value{
A named list containing the estimated 'inflection_point' and 'slope'.
}
\description{
Estimate true maturity parameters from ecosystem model input
}
