% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_truth.R
\name{get_truth_agecomp_monthly}
\alias{get_truth_agecomp_monthly}
\title{Calculate "true" monthly age composition for a species}
\usage{
get_truth_agecomp_monthly(truth_monthly)
}
\arguments{
\item{truth_monthly}{A tibble containing the monthly "truth" data for a species.
Returned from \code{\link[=get_truth_monthly]{get_truth_monthly()}}.}
}
\value{
A tibble containing the monthly age composition (e.g., biomass, catch,
and weight) for the specified species.
}
\description{
This function calculates the monthly age composition for a species
by selecting the relevant columns and filtering the data for the specified species.
}
\examples{
data(ewe_ecosim_base_nwatlantic)
data <- ewe_ecosim_base_nwatlantic
truth_agecomp_monthly <- ecosystemom:::get_truth_monthly(
  data,
  species_name = "menhaden",
  truth_type = "biomass"
) |>
  ecosystemom:::get_truth_agecomp_monthly()
}
\keyword{get_truth}
