% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_truth.R
\name{get_truth_agecomp_yearly}
\alias{get_truth_agecomp_yearly}
\title{Calculate "true" yearly age composition for a species}
\usage{
get_truth_agecomp_yearly(truth_agecomp_monthly)
}
\arguments{
\item{truth_agecomp_monthly}{A tibble containing the monthly age composition
data for a species. Returned from \code{\link[=get_truth_agecomp_monthly]{get_truth_agecomp_monthly()}}.}
}
\value{
A tibble containing the yearly age composition (e.g., biomass, catch,
and weight) for the specified species.
}
\description{
This function calculates the yearly age composition for a species
by averaging the monthly age composition values.
}
\examples{
data(ewe_ecosim_base_nwatlantic)
data <- ewe_ecosim_base_nwatlantic
truth_agecomp_yearly <- ecosystemom:::get_truth_monthly(
  data,
  species_name = "menhaden",
  truth_type = "biomass"
) |>
  ecosystemom:::get_truth_agecomp_monthly() |>
  ecosystemom:::get_truth_agecomp_yearly()
}
\keyword{get_truth}
