% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_truth.R
\name{get_truth_monthly}
\alias{get_truth_monthly}
\title{Calculate monthly "truth" for a species}
\usage{
get_truth_monthly(data, species_name, truth_type)
}
\arguments{
\item{data}{A tibble containing ecosystem model outputs from \code{load_model()}.}

\item{species_name}{A string specifying the species name to subset.}

\item{truth_type}{A string specifying the type of "truth" to calculate.}
}
\value{
A tibble containing the monthly "truth" data for the specified species.
}
\description{
This function calculates the monthly "truth" for a species by
filtering the data for the specified species and type
}
\examples{
data(ewe_ecosim_base_nwatlantic)
data <- ewe_ecosim_base_nwatlantic
truth_monthly <- ecosystemom:::get_truth_monthly(
  data,
  species_name = "menhaden",
  truth_type = "biomass"
)
}
\keyword{get_truth}
