% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_.R
\name{load_csv_environmental_data}
\alias{load_csv_environmental_data}
\title{Load environmental data from a CSV file}
\usage{
load_csv_environmental_data(file_path, lag_months, impacted_group)
}
\arguments{
\item{file_path}{A string. Path to the CSV file containing the environmental data.
The CSV file must contain the following columns:
\itemize{
\item \code{index}: Name of the environmental index (character).
\item \code{year}: Year corresponding to each value (integer).
\item \code{month}: Month corresponding to each value (integer).
\item \code{value}: Value of the index (numeric).
\item \code{unit}: Unit of the index (character).
}}

\item{lag_months}{An integer. The lag between environmental index and functional group, in months.
For example, if lag_months = 12, the environmental effect is assumed to influence the functional
group with a 12-month delay.}

\item{impacted_group}{A string indicating the functional group impacted by the environmental index.}
}
\value{
A tibble containing the environmental data with lag and functional group information.
}
\description{
Load environmental data from a CSV file
}
\examples{
data <- load_csv_environmental_data(
  file_path = file.path(
    system.file("extdata", package = "ecosystemom"),
    "ewe_ecosim_with_environmental_data_nwatlantic", "amo_lag1.csv"
  ),
  lag_months = 12,
  impacted_group = "menhaden 0"
)

}
