% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_.R
\name{load_csv_ewe}
\alias{load_csv_ewe}
\title{Load in EwE monthly output data}
\usage{
load_csv_ewe(file_path, model_years, functional_groups, verbose = TRUE)
}
\arguments{
\item{file_path}{The path to the CSV file containing the EwE output data.}

\item{model_years}{A numeric vector of years corresponding to the model run.
TODO: This could be derived automatically from an annual data file in the future.}

\item{functional_groups}{A vector of names of the functional groups in the model.}

\item{verbose}{Logical. If \code{TRUE} (default), prints informative messages.}
}
\description{
This function loads and standardizes output data from Ecopath with Ecosim (EwE) models.
}
\examples{
\dontrun{
# The following example is not run by default because these files are only
# included in the GitHub clone of the repository and not in the package data
load_csv_ewe(
  file_path = fs::path(
    "data-raw", "ewe_nwatlantic", "base_run", "biomass_monthly.csv"
  ),
  model_years = 1985:2017,
  functional_groups = get_functional_groups(
    file_path = fs::path(
      base_run_dir, "basic_estimates.csv"
    )
  )
)
}
}
