% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_gamma.R
\name{sample_gamma}
\alias{sample_gamma}
\title{Sample using a gamma distribution}
\usage{
sample_gamma(x, sd = 1)
}
\arguments{
\item{x}{A numeric vector of observations to sample from. If a single value
is provided, it will be recycled to the length of \code{sd}. Otherwise, it must
be the same length as \code{sd}.}

\item{sd}{A numeric vector of standard deviations to use in \code{\link[=rgamma]{rgamma()}}. If a
single value is provided, it will be recycled to the length of \code{x}.
Otherwise, it must be the same length as \code{x}.}
}
\value{
A numeric vector of the same length as \code{x} is returned.
}
\description{
Sample from a gamma distribution using the expected value in \code{x} and the
standard deviation in \code{sd}. The gamma shape and scale parameters are
derived from the input mean and standard deviation.
}
\details{
The gamma distribution is parameterized by shape and scale. Given an
expected value \code{x} and a standard deviation \code{sd}, the shape parameter is
calculated as \code{x^2 / sd^2} and the scale parameter as \code{sd^2 / x}.
}
\examples{
set.seed(123)
sample_gamma(1, sd = 1)
set.seed(123)
sample_gamma(1, sd = c(0, 1))
sample_gamma(1:5, sd = 1)

}
