% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_lognormal.R
\name{sample_lognormal}
\alias{sample_lognormal}
\title{Sample using a lognormal distribution}
\usage{
sample_lognormal(x, sd = 1)
}
\arguments{
\item{x}{A numeric vector of observations to sample from. These observations
should be on the normal scale. If a single value is provided, it will be
recycled to the length of \code{sd}. Otherwise, it must be the same length as
\code{sd}.}

\item{sd}{A numeric vector of standard deviations to use in \code{\link[=rnorm]{rnorm()}}. If a
single value is provided, it will be recycled to the length of \code{x}.
Otherwise, it must be the same length as \code{x}.}
}
\description{
Sample from a standard normal in log space and apply this
exponentiated observation error to the observations, i.e, \code{x}, after the
simulated error values are bias corrected.
}
\details{
\subsection{Historical methods}{
\subsection{{atlantisom}}{

A lognormal distribution was used to sample using a mean of the natural log
of the expected value minus the input variance divided by two and the input
standard deviation. Where the input variance was calculated as the natural
log of the coefficient of variation squared plus one.
}

\subsection{IP4EBFM}{

A normal distribution was used to sample the error using a mean of zero and
a standard deviation provided by the input. This error was then
exponentiated but not bias corrected before multiplying it by the
observation.
}

\subsection{{ss3sim}}{

A normal distribution was used to sample the error using a mean of zero and
a standard deviation provided by the input. This error was then bias
corrected and exponentiated before multiplying it by the observation.
}

}
}
\examples{
# Note that setting the seed on a vectorized implementation of this function
# will not lead to the same sample being returned for each element of x
# because a single random number is generated for the entire vector.
set.seed(123)
sample_lognormal(1:10, sd = 1)
set.seed(123)
sample_lognormal(1, sd = 1)
sample_lognormal(2, sd = 1)

# You can also use the function with a vector of standard deviations
set.seed(123)
sample_lognormal(1:10, sd = 1:10)
}
