% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_functional_groups.R
\name{get_functional_groups}
\alias{get_functional_groups}
\title{Get the functional groups from an EwE file}
\usage{
get_functional_groups(file_path)
}
\arguments{
\item{file_path}{The path to the EwE file.}
}
\value{
A tibble with the following columns:
\itemize{
\item function_group. The full functional group name from the model. This
column is helpful for error checking and debugging.
\item species. The species name from the functional group name. This
column should only contain text strings and no numbers or special
characters. There will potentially be multiple rows with the same
species name because there can be multiple age groups for a given
species.
\item group. The group name as a string. This column can contain digits,
special characters, and text strings. It is used to
delineate the group within a species. Not all species will have
multiple groups.
\item functional_group_snake_case. A standardized version of the functional
group name. The functional_group is converted to lowercase and swaps
spaces or hyphens for underscores.
}
}
\description{
Functional group names are useful for the column names of output data from
an EwE model, and thus, this function is a way to get them automatically
from the EwE output. The functional group names come from the basic
estimates file.
}
\examples{
get_functional_groups(
  file_path = fs::path(
    system.file("extdata", package = "ecosystemom"),
    "ewe_ecosim_with_environmental_data_nwatlantic", "basic_estimates.csv"
  )
)
}
